refactor(law-model): extract canonical law document model into a leaf crate
Introduce regelrecht-law-model, a dependency-light leaf crate (serde only)
holding the single Rust representation of the law-YAML document model: the
document structs/enums previously in engine's article.rs plus the
Value/Operation/ParameterType types from engine's types.rs.

The engine re-exports everything at the historical engine::article and
engine::types paths, so no downstream code changes. YAML loading with the
security limits (size/array caps, RFC-013 schema-version + content-hash) stays
in the engine as the new LawLoad extension trait, since it depends on the
engine's config and error type; call sites import the trait.

The JSON schema remains the source of truth (RFC-013); no codegen is
introduced. This is the foundation for retiring the partial model copies in
corpus, harvester, tui and pipeline.
