+++ packages/Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "editor-api",
     "engine",
     "harvester",
+    "law-model",
     "pipeline",
     "shared",
     "tui",
