+++ packages/engine/Cargo.toml
@@ -18,6 +18,7 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 regelrecht-shared = { path = "../shared" }
+regelrecht-law-model = { path = "../law-model" }
 serde.workspace = true
 serde_yaml_ng.workspace = true
 serde_json.workspace = true
