+++ packages/engine/benches/law_loading.rs
@@ -1,5 +1,5 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use regelrecht_engine::ArticleBasedLaw;
+use regelrecht_engine::{ArticleBasedLaw, LawLoad};
 use std::path::PathBuf;
 
 fn corpus_path() -> PathBuf {
