+++ packages/law-model/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "regelrecht-law-model"
+version = "0.1.0"
+edition = "2021"
+authors = ["MinBZK <noreply@minbzk.nl>"]
+description = "RegelRecht canonical law-YAML document model - the single Rust representation of the law format"
+license = "EUPL-1.2"
+repository = "https://github.com/MinBZK/regelrecht"
+homepage = "https://github.com/MinBZK/regelrecht"
+keywords = ["law", "legal", "dutch", "regelrecht", "types"]
+categories = ["parsing"]
+
+[dependencies]
+regelrecht-shared = { path = "../shared" }
+serde.workspace = true
+serde_json.workspace = true
+
+[dev-dependencies]
+serde_json.workspace = true
+
+[lints]
+workspace = true
